requirements.txt
absl-py
amqp
anyascii
arabic-reshaper
asgiref
asn1crypto
astroid
astunparse
attrs
autopep8
backports.entry-points-selectable
beautifulsoup4
billiard
cachetools
celery
certifi
cffi
chardet
charset-normalizer
click
click-didyoumean
click-plugins
click-repl
colorama
comtypes
contourpy
crispy-bootstrap5
cron-descriptor
cryptography
cssselect2
cycler
defusedxml
diagram
diff-match-patch
dill
distlib
dj-database-url
Django
django-bootstrap-v5
django-bootstrap4
django-celery-beat
django-cors-headers
django-cors-middleware
django-crispy-forms
django-email
django-environ
django-filter
django-import-export
django-modelcluster
django-taggit
django-timezone-field
django-treebeard
django-widget-tweaks
djangorestframework
docopt
draftjs-exporter
EasyProcess
entrypoint2
et-xmlfile
face-recognition-models
filelock
Flask
flatbuffers
fonttools
freetype-py
gast
ghs
google-auth
google-auth-oauthlib
google-pasta
graphviz
grpcio
gunicorn
h5py
halo
html5lib
idna
importlib-metadata
iniconfig
isort
itsdangerous
Jinja2
joblib
keras
Keras-Preprocessing
keyboard
kiwisolver
kombu
l18n
lazy-object-proxy
libclang
log-symbols
lxml
Markdown
MarkupPy
MarkupSafe
matplotlib
mccabe
MouseInfo
mss
notify2
numpy
oauthlib
odfpy
opencv-python
openpyxl
opt-einsum
oscrypto
packaging
pandas
Pillow
pipenv
pipreqs
platformdirs
pluggy
ppt2pdf
prompt-toolkit
protobuf
psycopg2
py
pyasn1
pyasn1-modules
PyAutoGUI
pycairo
pycodestyle
pycparser
pyfiglet
PyGetWindow
Pygments
pyHanko
pyhanko-certvalidator
PyInquirer
pyjokes
pylint
PyMsgBox
pyparsing
pypdf
PyPDF2
pyperclip
PyRect
pyscreenshot
PyScreeze
pytest
python-bidi
python-crontab
python-dateutil
pytweening
pytz
PyYAML
qrcode
regex
reportlab
requests
requests-oauthlib
retry-requests
rlPyCairo
rsa
scikit-learn
scipy
six
soupsieve
spinners
sqlparse
sslcommerz-lib
stripe
svglib
tablib
telepath
tensorboard
tensorboard-data-server
tensorboard-plugin-wit
tensorflow
tensorflow-estimator
tensorflow-io-gcs-filesystem
termcolor
threadpoolctl
tinycss2
tk
toml
tomli
tomlkit
topsis-jamesfallon
TOPSIS-Yash-101803064
typed-ast
typing_extensions
tzdata
tzlocal
uritools
urllib3
vine
virtualenv
virtualenv-clone
wagtail
wcwidth
webencodings
Werkzeug
whitenoise
Willow
wrapt
xhtml2pdf
xlrd
XlsxWriter
xlwt
yarg
zipp
cloudinary
django-cloudinary-storage
pytz
